Changed enum to enum class in windowParams. Changed Window class and base_project to work with enum class. Disabled numberOfWindows in WMS. Initilaised width and height in windowParams. Applied static cast to window.size() in getWindowCount(). This is to supress a warning; size limits are already in place so the warning should not cause any issues.

+++ Services/WindowManagementService.cpp
@@ -90,5 +90,6 @@ void WindowManagementService::closeAllWindows() {
 
 // return number of windows
 int WindowManagementService::getWindowCount() {
-	return windows.size();
+	return static_cast<int>(windows.size());
+	// warning, size_t to int; possible loss of data :: there is already a limit in place, and that limit is not expected to be high enough to cause any issues
 }
